# Enhancement/Improve performance of audio visualizer generation